+++ database/migrations/2020_02_02_203230_create_order_details_table.php
@@ -14,8 +14,8 @@ class CreateOrderDetailsTable extends Migration
     public function up()
     {
         Schema::create('order_details', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('product_id')->unsigned();
+            $table->increments('id');
+            $table->integer('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products');
             $table->decimal('price',8,2);
             //$table->string('status');
@@ -23,7 +23,7 @@ class CreateOrderDetailsTable extends Migration
             $table->integer('quantity_products');
             $table->date('estimated_delivery');
             $table->date('delivery_date');
-            $table->bigInteger('order_id')->unsigned();
+            $table->integer('order_id')->unsigned();
             $table->foreign('order_id')->references('id')->on('orders');
             $table->timestamps();
         });
